requirements.txt
pytest
requests
python-dotenv
jsonschema
hypothesis
